pyproject.toml
[project]
name = "oaisandbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.3.1",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
]
